Change type of product price

+++ database/migrations/2023_02_02_104512_create_products_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('name');
             $table->string('description')->nullable();
             $table->integer('status');
-            $table->decimal('price')->nullable();
+            $table->float('price')->nullable();
             $table->timestamps();
             $table->softDeletes();
